refacored themes
